refactor(admin): extract block-form field-dom-sync module
Verbatim move of the in-place field DOM patchers (updateImageFieldDom,
updateFileFieldDom, seedAltInput, seedCaptionInput) into
block-form/field-dom-sync.ts. These are the helpers behind the
no-re-render-on-value-change convention; both the picker selection path
and the mount wiring keep calling them unchanged.

Part of #38, slice 3/6.
