+++ src/routes/admin/client/block-form.ts
@@ -78,6 +78,12 @@ import {
   withLocaleHint,
 } from './block-form/field-helpers.js';
 import { computeEffectiveAccept } from './block-form/file-accept.js';
+import {
+  seedAltInput,
+  seedCaptionInput,
+  updateFileFieldDom,
+  updateImageFieldDom,
+} from './block-form/field-dom-sync.js';
 
 export { checkArrayLimitReached };
 export type { ArrayLimitInfo };
@@ -154,98 +160,6 @@ function closePickerDialog(): void {
     triggerBtn?.focus();
   }
   activePickerInputId = null;
-}
-
-/**
- * Update an image field's visible DOM IN PLACE to reflect `url` (empty string =
- * cleared). Fields are not re-rendered on value change, so the preview, filename,
- * "selected" state, and Choose/Clear labels must be mutated directly here.
- *
- * Pure presentation: it does NOT touch the hidden input value, dispatch events,
- * or alter the value contract — callers own that.
- */
-function updateImageFieldDom(hiddenInput: HTMLInputElement, url: string): void {
-  const field = hiddenInput.closest<HTMLElement>('.cms-image-field');
-  if (!field) return;
-  const hasValue = url.length > 0;
-  const filename = hasValue ? imageFilenameFromUrl(url) : '';
-  field.classList.toggle('cms-image-field--has-value', hasValue);
-
-  const previewWrap = field.querySelector<HTMLElement>('[data-image-preview]');
-  if (previewWrap) {
-    previewWrap.innerHTML = hasValue
-      ? `<img src="${escapeAttr(url)}" alt="${escapeAttr(filename)}" class="cms-image-field-thumb" data-image-thumb>`
-      : `<span class="cms-image-field-placeholder" aria-hidden="true">${imagePickerIconSvg}</span>`;
-  }
-
-  const nameEl = field.querySelector<HTMLElement>('.cms-image-field-name');
-  if (nameEl) {
-    nameEl.classList.toggle('cms-image-field-name--empty', !hasValue);
-    nameEl.textContent = hasValue ? filename : 'No image selected';
-    if (hasValue) nameEl.setAttribute('title', filename);
-    else nameEl.removeAttribute('title');
-  }
-
-  const chooseLabel = field.querySelector<HTMLElement>('[data-choose-label]');
-  if (chooseLabel) chooseLabel.textContent = hasValue ? 'Replace' : 'Choose image';
-}
-
-/**
- * Update a file field's visible DOM IN PLACE to reflect the new value.
- * Mirror of updateImageFieldDom for the 'file' prop type.
- */
-function updateFileFieldDom(hiddenInput: HTMLInputElement, value: FileFieldValue): void {
-  const field = hiddenInput.closest<HTMLElement>('.cms-file-field');
-  if (!field) return;
-  const hasValue = !isEmptyFileValue(value);
-  const displayName = hasValue ? (value.filename ?? imageFilenameFromUrl(value.url)) : '';
-  field.classList.toggle('cms-file-field--has-value', hasValue);
-
-  const nameEl = field.querySelector<HTMLElement>('.cms-file-field-name');
-  if (nameEl) {
-    nameEl.classList.toggle('cms-file-field-name--empty', !hasValue);
-    nameEl.textContent = hasValue ? displayName : 'No file selected';
-    if (hasValue) nameEl.setAttribute('title', displayName);
-    else nameEl.removeAttribute('title');
-  }
-
-  const chooseLabel = field.querySelector<HTMLElement>('[data-file-choose-label]');
-  if (chooseLabel) chooseLabel.textContent = hasValue ? 'Replace' : 'Choose file';
-}
-
-/**
- * Seed the visible alt-override input for an image field from the current hidden-input JSON.
- * Reads parseImageValue(hidden.value).alt and sets the alt input value.
- */
-function seedAltInput(hiddenInput: HTMLInputElement, altOverride?: string): void {
-  const field = hiddenInput.closest<HTMLElement>('.cms-image-field');
-  if (!field) return;
-  const altInput = field.querySelector<HTMLInputElement>('[data-image-alt-for]');
-  if (!altInput) return;
-  if (altOverride !== undefined) {
-    altInput.value = altOverride;
-  } else {
-    const current = parseImageValue(hiddenInput.value);
-    altInput.value = current.alt ?? '';
-  }
-}
-
-/**
- * Seed the visible caption input for an image field from the current hidden-input JSON.
- * Reads parseImageValue(hidden.value).caption and sets the caption input value.
- * If captionOverride is provided, that value is used directly (used on picker pick and clear).
- */
-function seedCaptionInput(hiddenInput: HTMLInputElement, captionOverride?: string): void {
-  const field = hiddenInput.closest<HTMLElement>('.cms-image-field');
-  if (!field) return;
-  const captionInput = field.querySelector<HTMLInputElement>('[data-image-caption-for]');
-  if (!captionInput) return;
-  if (captionOverride !== undefined) {
-    captionInput.value = captionOverride;
-  } else {
-    const current = parseImageValue(hiddenInput.value);
-    captionInput.value = current.caption ?? '';
-  }
 }
 
 function selectPickerImage(value: ImageFieldValue): void {
